server/requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.1.0
fastapi==0.90.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
install==1.3.5
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pydantic==1.10.4
pytest==7.2.1
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.23.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
zipp==3.12.1
numpy~=1.21.6
opencv-python~=4.7.0.68
easyocr~=1.6.2